Band convergence reached for 50bands (v 18 c 32) till 12.5 eV
The convergence was reached for a total of 50 bands that includes 18 valence
bands and 32 conduction bands. The convergergence was completed till 12.5ev
corresponding almost to the limit between near and far UV radiation. The
convergence was done to this point for the linear response; other responses
are converged till 15eV.